Fix compilation on win64
TBR=bmeurer@chromium.org
LOG=n
BUG=none

Review URL: https://codereview.chromium.org/613093004

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@24440 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ test/cctest/compiler/test-operator.cc
@@ -29,10 +29,10 @@ TEST(TestOperatorMnemonic) {
 
 TEST(TestSimpleOperatorHash) {
   SimpleOperator op1(17, Operator::kNoProperties, 0, 0, "Another");
-  CHECK_EQ(17, op1.HashCode());
+  CHECK_EQ(17, static_cast<int>(op1.HashCode()));
 
   SimpleOperator op2(18, Operator::kNoProperties, 0, 0, "Falsch");
-  CHECK_EQ(18, op2.HashCode());
+  CHECK_EQ(18, static_cast<int>(op2.HashCode()));
 }
